# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI 코딩 어시스턴트 API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  conversationId: number;
  /** @nullable */
  model?: string | null;
}

export interface MessageResponse {
  id: number;
  content: string;
  role: string;
  conversationId: number;
  createdAt: string;
}

export interface Message {
  id: number;
  content: string;
  role: string;
  conversationId: number;
  createdAt: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  messageCount?: number | null;
}

export interface ConversationDetail {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export interface ChatStats {
  totalConversations: number;
  totalMessages: number;
  todayMessages: number;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
};