# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI 코딩 어시스턴트 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/send:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
        "500":
          description: Server error
...
components:
...
  schemas:
...
    MessageInput:
      type: object
      required: [content, conversationId]
      properties:
        content:
          type: string
          minLength: 1
        conversationId:
          type: integer
        model:
          type: ["string", "null"]
    MessageResponse:
      type: object
      required: [id, content, role, conversationId, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        role:
          type: string
        conversationId:
          type: integer
        createdAt:
          type: string
...