# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI 코딩 어시스턴트 API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResponse {
  id: number;
  content: string;
  role: string;
  conversationId: number;
  createdAt: string;
}